+++ SistemaRestaurante/src/negocio/ComandaABM.java
@@ -12,7 +12,7 @@ public class ComandaABM {
 	private ComandaDao dao = new ComandaDao();
 
 	public int agregarComanda(GregorianCalendar fecha, Mesa mesa, Cliente cliente,
-			Camarero camarero, int numHabitacion, boolean activo){
+			Camarero camarero, long numHabitacion, boolean activo){
 		Comanda c = new Comanda (fecha, mesa, cliente, camarero,
 				numHabitacion, activo);
 		return dao.agregarComanda(c);
